Fix problem with data type of resolver.source
IP addresses are the only thing that makes sense here, so arbitary
string shouldn't be allowed.
Also allowing undef is problematic since get() returns undef for unset
properties.
The caller should be able to determine whether or not the property is
set or unset using the return value of get().