Avoid returning self
The original purpose of returning self was to facilitate the method
chaining builder pattern that's popular in some languages.

    my $profile = Zonemaster::Engine::Profile->new
        ->set( 'net.ipv4', 1 )
        ->set( 'net.ipv6', 0 )
        ->merge( $other );

The down side is that it could fool people into believing that a modified copy
is returned, when in fact it is self that is modified and returned.

    Zonemaster::Engine::Profile->effective->merge( $default->set( 'no_network', 0 ) );

However it turns out perltidy doesn't play nice with the method
chaining so its value is greatly diminished and only the down sides
remain. In the end it's better to not return anything from set() and
merge().