Remove workaround for previously unsupported C locale
The develop versions of both Engine and CLI use the proper English
message strings for message ids, so neither of those need this
workaround. Older CLI versions don't need it either because the English
message ids and message strings have been identical for a very long
time. Backend is not affected because it doesn't have any translations
of its own. Any other clients are assumed to use proper English in their
message ids, so this workaround isn't important for those either.