Relieve the PO files from being make targets
The fact that the PO files depend on themselves together with the need
to sometimes NOT update them isn't a good fit for the Makefile model.

Currently we're trying to shoehorn the PO files into being good make
targets by adding the touch-po target under share/ and having "make all"
at the top level run "make touch-po all" under share/ (instead of just
"make all").

A better solution, as presented here, is to relieve the PO files from
being make targets and instead relying on the share/update-po script to
update the msgids only when appropriate.