Include perl-mailtools in Docker image
The Docker image for zonemaster-engine was broken, because the
Mail::Address module, required by Email::Valid, was missing from it.

The missing module is found in the perl-mailtools package on Alpine
Linux, so this commit adds perl-mailtools to the list of packages to be
installed in the second stage image.