Dockerfile: add build dependencies for Net::IP::XS
It turns out it was insufficient to just add Net::IP::XS to the list of
packages to install in the Dockerfile. Because it’s the first XS package
to be installed, it also requires GCC and the C header files for both
Perl (for EXTERN.h) and musl libc (for sys/types.h).