Align unit tests for ZONE11 test with #1297
Change the format of the test declarations inside t/Test-zone11.t so
that it conforms to the format specified and agreed to in #1297.

This is unfortunately not a lossless conversion: the old format made it
possible to specify extra checks such as “for zone
all-different-spf.zone11.xa, the Z11_DIFFERENT_SPF_POLICIES_FOUND should
be output 3 times”. Hopefully this can be brought back somehow later.

Also, for some reason, the changes in the .t file required a new test
run with ZONEMASTER_RECORD=1 set in the environment.