Dockerfile: also install Mail::SPF in 2nd stage
Installing Mail::SPF through the apk command caused Net::DNS to be
installed by the apk package manager, instead of cpanm.

But this package was missing in the second stage. This caused
Zonemaster::Engine to fail to load because Net::DNS was missing.
However, the underlying problem really was that Mail::SPF was not
installed in the second change.

Because Mail::SPF is a runtime dependency for Zonemaster::Engine, we
need to ensure that this module is installed in both stages.