Fix polymorphism in Translator
In Zonemaster::Backend, the codebase defines a
Zonemaster::Backend::Translator class which inherits from
Zonemaster::Engine::Translator, because it needs to override the
translate_tag() method with a custom version.

However, one change introduced in zonemaster-engine#1319 caused this
inheritance to no longer work properly.

The main symptom was that commands such as `zmb get_test_results`, that
rely on translating messages pulled from the database, stopped working.

The root cause of the problem is that when calling
Zonemaster::Backend::Translator->new(), the newly-created instance was
always blessed as Zonemaster::Engine::Translator (the superclass), not
as the inheriting class. Thus, the translate_tag() method was always
that of the superclass instead of the inheriting class. Hence the crash.

Fixing this is easy, fortunately.