Translator: fix instance() method, documentation
A programming error in Zonemaster::Engine::Translator’s instance()
method caused the global variable $instance to be immediately undefined
right after the initialize() function set it to an instance of a
Translator object, if that singleton object hadn’t been previously
instanciated.

The documentation for Zonemaster::Engine::Translator’s initialize()
method was also incorrect: this function takes a raw list of arguments,
not a hash reference.

Do not use the new() method in the corresponding unit tests anymore,
because that method is deprecated.