$p masking in Zonemaster::Engine::Nameserver
This change fixes the following warning caught by Debian Perl Team's
autopkgtest starting with Zonemaster::Engine 5.0.0:

	"my" variable $p masks earlier declaration in same scope at /usr/share/perl5/Zonemaster/Engine/Nameserver.pm line 363.

This warning occurs because the variable is declared multiple times
behind "if" statements, and another time unconditionally later in the
subroutine.

This patch proceeds to the $p variable declaration earlier in the
subroutine unconditionnally, so they are always guaranteed to be
available and unmasked.  The only drawback is that the $in_cache
variable now needs to be declared in a separate statement.

Signed-off-by: Étienne Mollier <emollier@debian.org>