Update unit tests and unit tests data
Use a different query name for each query as to make sure that we do not get a cached response from previous, equivalent queries.
This is needed because the tested object is actually the underlying object in the "dns" attribute of the Zonemaster::Engine::Nameserver class,
which is only updated by the "_query()" method, and not the "query()" method.