DNSName.pm: trim some fat off constructor
Simplify the constructor, which comes with a considerable speedup (about
14% less time spent constructing new objects).

Function calls are expensive in Perl, so we start by eliminating the
_new() method and replacing them with direct calls to
Class::Accessor::new(). Also, there is no need to do something like

  my $class = ref $proto || $proto

because Class::Accessor::new() already does that for us.

We trade off a little bit of safety for speed as well: the new()
function is now assuming that it’s always called with at most 1
parameter. We also remove a useless type check because the error case
was unreachable.

Given how often new objects are constructed, speeding this function up
even slightly has a significant effect on Zonemaster’s run time.