MethodsV2: memoize get_parent_ns_ips()
This method is used in a few places, including in other methods in
MethodsV2. As such, it may be called repeatedly with identical arguments
in different places. It is expensive in time due to the number of
requests that this method generates, so it is particularly interesting
to memoize. Doing so yields a very significant performance boost.