DNSName.pm: optimize str_cmp()
A few modifications are made to str_cmp, which is one of the most-used
functions in Zonemaster::Engine.

Firstly, not unpacking @_ gives us a first speed boost.

Secondly, test whether we are comparing with another
Zonemaster::Engine::DNSName object (which is almost always the case), so
we can use a shortcut if that is the case.

Thirdly, use a more efficient regular expression to remove a trailing
dot from a name except if it is the only character in the string. The
trick, which works with Perl 5.10.0 and up, uses the relatively obscure
\K token in the regular expression, which works more or less like a
variable-length lookbehind.