DNSName.pm: apply a few more optimizations
I’m less sure about whether these particular optimizations do any good,
because the speedup I’ve measured is too small to be statistically
significant. The code in question is not used often.

But because we know that the “name” attribute of a
Zonemaster::Engine::Zone is always a Zonemaster::Engine::DNSName, we can
take a little shortcut.