Add failing test for Address03; disable stale test
Add a proper unit test for the Address03 test case. Currently, it fails
because of PTR-IS-GOOD-CNAME-2, because there is a bug in this scenario,
but we’re going to fix it in a follow-up commit.

A stale test, involving torsas.se which formerly had classless
in-addr.arpa delegation, is disabled, because it isn’t classless
in-addr.arpa anymore.