Implement a DSL for unit tests
There is a proto-DSL already, in t/TestUtil.pm, that helps us test
Zonemaster test cases with respect to predefined scenarios. But we need
to go further.

Scenarios are defined in terms of a zone name (which names a zone that
is configured in a predetermined way), some possible extra configuration
for the test case (such as “fake NS” and “fake DS” records), and a list
of messages that we expect Zonemaster to emit when testing the zone
name.

Currently, when using t/TestUtil.pm, we list scenarios in a array of
arrays. It has been working fine for now, but it’s lacking a feature:
testing message arguments in addition to message tags. How do we specify
predicates that message *arguments* should satisfy? This is where the
current method shows its limit.

The DSL is meant to be easy to read, not only for those who routinely
write such unit tests, but also for more casual readers. Its system of
keywords leaves room for extensibility when the need for new features
arises. It provides shorthands that aid writing unit tests that look
similar, without compromising readability. And last but not least, the
syntax allows for specifying properties that message arguments are to
satisfy when looking for messages of a certain tag (which isn’t fully
implemented yet).

Care has been taken to retain maximum compatibility with the features of
TestUtil.pm’s perform_testcase_testing() function in most areas.

In addition for its improved facilities for testing message arguments,
the DSL introduces another extension: a distinction between “todo”
tests (i.e. not testable because of a bug in the implementation, but
otherwise the scenario is fine) and “not_testable” tests (i.e. not
testable because the scenario is not implemented or cannot be
implemented with current means).