Port t/Test-basic02.t to DSL
Rewriting this test file to use the DSL helped me find a typo in the
original t/Test-basic02.t file that caused the GOOD-2 scenario to point
to good-1.basic02.xa instead of good-2.basic02.xa (see PR #1465).

It’s a good demonstration of why it can be important to design data
structures and formats in such a way that internal repetition is avoided
unless absolutely necessary: when done right, it makes subtle errors
like these more obvious.