t/Test-basic02.t: fix typo in zone name
The scenario named GOOD-2 erroneously referred to a test zone named
good-1.basic02.xa, instead of good-2.basic02.xa. Because of that,
scenario GOOD-2 was not properly tested.

Simply fixing the typo was not enough. No data was recorded about
good-2.basic02.xa at all, so the companion .data file also needed to be
regenerated.