Fix and enable unit test in Test-consistency06.t
Scenario MULT-SOA-MNAMES-NO-DEL-UNDEL-2 was marked as not testable, but
it turned out that it was due to errors both in the test scenario
implementation (i.e. the zone files and CoreDNS configuration: see
zonemaster/zonemaster#1414) and a small typo in the .t file.

In the .t file, scenario MULT-SOA-MNAMES-NO-DEL-UNDEL-2 erroneously
listed ns3.mult-soa-mnames-no-del-undel-2.consistency06.xb twice in its
undelegated data. The second instance should be ns4, not ns3.

After fixing these problems, we can enable the scenario in
Zonemaster-Engine’s test suite and rerecord the test data.