Fix undelegated data in Consistency06 test scenario
Scenario MULT-SOA-MNAMES-NO-DEL-UNDEL-2 erroneously listed
ns3.mult-soa-mnames-no-del-undel-2.consistency06.xb twice in its
undelegated data. The second instance should be ns4, not ns3.

Unfortunately, that doesn’t fix the unit test because of a problem in
the implementation of Consistency06. Therefore, the scenario
MULT-SOA-MNAMES-NO-DEL-UNDEL-2 stays disabled in the unit tests.

Nevertheless, I’ve rerecorded the unit test data in the hopes that it
will not be needed again if someone attempts to fix the test case.