Implement testing of message arguments
An issue, zonemaster/zonemaster#1401, proposes to “[u]pdate a test
scenario with requirements for message arguments”.

While as of writing this commit for the first time (2025-07-30), no
scenario has been updated yet, the tooling is implemented.

Message arguments can be tested as it is specified in the DSL’s
documentation. If such a test fails, we try our best to print as helpful
a message as possible: the test prints what exactly it was looking for
and which messages it has examined.

This debugging output could be improved by explaining why a given
message has failed a match, but the current level of detail is hopefully
plenty enough for debugging purposes.

Below is a sample TAP output detailing a failure of such a test. Long
lines are folded in order not to exceed 78 columns (the offendings lines
end with a backslash):

    t/Test-basic02.t .. 10/?
        #   Failed test 'Messages of tag 'B02_AUTH_RESPONSE_SOA' exist with\
        #   specified arguments'
        #   at t/Test-basic02.t line 112.
        # Looked for a message whose tag is 'B02_AUTH_RESPONSE_SOA'
        #   where 'domain' equals 'mixed-1.basic02.xa'
        #     and 'ns_list' equals 'ns1.mixed-1.basic02.xa/127.12.2.31;\
        #     ns1.mixed-1.basic02.xa/fda1:b2:c3:0:127:12:2:31'
        #     and contains no argument other than those listed above
        # Here are all messages that unsuccessfully matched:
        #   B02_AUTH_RESPONSE_SOA domain="mixed-1.basic02.xa"; \
        #     ns_list=ns1.mixed-1.basic02.xa/127.12.2.31;\
        #     ns1.mixed-1.basic02.xa/fda1:b2:c3:0:127:12:2:31
    #   Failed scenario 'MIXED-1'
    #   at t/Test-basic02.t line 117.
    # Looks like you failed 1 test of 26.
    t/Test-basic02.t .. Dubious, test returned 1 (wstat 256, 0x100)
    Failed 1/26 subtests