Compute correct cache key for memoization
The get_parent_ns_ips() method in MethodsV2 takes a class (string) and a
Zonemaster::Engine::Zone object. It is memoized thanks to the facility
provided by the Memoize module.

However, it turns out that the cache key computed by Memoize is
incorrect. This could lead to suboptimal performance and in some cases
completely wrong results. In particular, it could cause the
t/methodsv2.t unit test to fail intermittently – which can be a real
nuisance with continuous integration.

By default, Memoize takes the string representations of each item in @_
joined by the ASCII 0x1c (FS, Field Separator) character. The $zone
argument is a blessed reference to Zonemaster::Engine::Zone, and the
package does not provide a stringification operator override; therefore
a nonsensical string like Zonemaster::Engine::Zone=HASH(0x62c9f89bc410)
is used as a cache key instead of the zone name.

This commit provides a custom normalizer to the call to memoize() so
that the zone name is used as part of the cache key. It seems to fix the
problem (I tested by running t/methodsv2.t 50 times in a row) and, as a
bonus, also accelerates the unit test suite somewhat. Maybe real-world
domain tests might also benefit from it?