Rerecord test data after updating MethodV2
Somehow, introducing Get-Parent-NS-Names-and-IPs, then redefining
Get-Parent-NS-IPs in terms of Get-Parent-NS-Names-and-IPs, has had side
effects on some unit tests that perform tests against live zones:
suddenly, they tried to perform queries to a variety of name
servers (among which M-root).

I haven’t fully investigated the reason why the tests failed, but it
might be because the order of the servers being queried has changed. I
also suspect some of the name servers have changed IP addresses since
recording.

The offending .t files are t/Test.t, t/Test-dnssec.t and
t/Test-connectivity.t.

Rerecording the corresponding data files is enough to fix everything,
except for t/Test-dnssec.data where it was necessary to splice part of
the newly recorded data into the existing data so that everything
continues to work. It is rather inelegant, but it gets the job done
until the affected tests are migrated to the new scenario-based
framework.