Update root hints after B-root renumbering
Update the built-in root hints file: this hasn’t been done in three
years, and b.root-servers.net’s IP addresses have changed since (see
[1]).

This is to make sure that we don’t send traffic to IP addresses
belonging to _former_ root servers (see for example [2] for the case of
J-root).

I think this also reveals an oversight in our release procedure: we
ought to make sure that every release of Zonemaster ships with a fresh
and up-to-date root hints file. I don’t think this is currently the
case.

Unit test data needs to be rerecorded as well. This is done except for
t/old-bugs.data; that file cannot be updated easily.

[1]: https://b.root-servers.org/news/2023/05/16/new-addresses.html
[2]: https://indico.dns-oarc.net/event/55/contributions/1176/attachments/1134/2372/wessels-oarc45-oldjroot.pdf